Explicitly cast some constants as floats in OpenCL code
Some compilers are treating 0.567 as double, while others as
float. Some break with 0.567f notation too. So explicit C casts are
needed to force floats and prevent compile time errors.

+++ vessel_integration.cl
@@ -46,9 +46,9 @@ inline float sqr(float n)
 
 inline float viscosityFactor(float D, float Hct)
 {
-	const float C = (0.8+exp(-0.075*D)) * ((1/(1 + 1e-11*pown(D, 12)))-1.0) + (1/(1+1e-11*pown(D,12)));
-	const float Mi45 = 220 * exp(-1.3*D) - 2.44*exp(-0.06*powr(D, 0.645)) + 3.2;
-	return (1.0 + (Mi45-1.0)*(powr(1.0-Hct, C)-1.0)/(powr(1.0-0.45, C)-1.0)) / 3.2;
+	const float C = (0.8+exp((float)-0.075*D)) * ((1/(1 + 1e-11*pown(D, 12)))-1.0) + (1/(1+1e-11*pown(D,12)));
+	const float Mi45 = 220 * exp((float)-1.3*D) - 2.44*exp(-0.06*powr(D, (float)0.645)) + 3.2;
+	return (1.0 + (Mi45-1.0)*(powr((float)1.0-Hct, C)-1.0)/(powr((float)1.0-(float)0.45, C)-1.0)) / 3.2;
 }
